Convert decoded tests to event parser
